# ARROW-2222: handle untrusted inputs