@crepererum 2018-03-24
OK, I'll provide a comparison later.

BTW: fuzzing just got easier w/ llvm/clang 6.0. Since the fuzzer is now officially a part of compiler-rt, it is likely that it's included in many distributions / packages by default. For example. Arch bundles a feature-complete compiler-rt, so no additional preparation is required to get fuzzers working (just tested it).